backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-cors-headers>=4.7,<5.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.1,<2.0
